2024.04.09 PlayerWeapon
Player weapon:
- Each character has its weapon.
- Functions like State Pattern.

+++ Assets/Scripts/Interface/IHit.cs
@@ -1,4 +1,4 @@
 public interface IHit
 {
-    public void Hit(byte damage, float direction);
+    public void Hit(ushort damage, sbyte direction);
 }
